FT connector: add whitelist tests
